+++ server.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const path = require('path');
+const cron = require('node-cron');
 const Discord = require('discord.js');
 require('discord-reply');
 const Storage = require('./util/Storage');
@@ -8,6 +9,7 @@ const config = require('./config.json');
 const tokenFilePath = 'TOKEN.txt';
 const ownerIdPath = 'OWNER_ID.txt';
 const commandDir = 'commands/';
+const cronDir = 'crontabs/';
 const validCommandRegex = /^[a-z0-9]+$/i;
 const defaultPrefix = config.defaultPrefix; // intended to be an ascii representation of a basic tank
 
@@ -16,6 +18,7 @@ const storage = new Storage();
 const token = fs.existsSync(tokenFilePath) ? fs.readFileSync(path.join(__dirname, tokenFilePath), 'utf8').trim() : false;
 const ownerId = fs.existsSync(ownerIdPath) ? fs.readFileSync(path.join(__dirname, ownerIdPath), 'utf8').trim() : false;
 const commands = Object.create(null);
+const cronsToRunImmediately = [];
 let viewOwnerAsAdmin = false;
 
 if (!token) {
@@ -31,11 +34,11 @@ for (const eachModuleFile of fs.readdirSync(commandDir)) {
     if (!validCommandRegex.test(commandName)) {
       throw new Error(`Invalid command name ${commandName}. Must match regex ${validCommandRegex}`);
     }
-    if (['help'].includes(commandName)) {
+    if (['help', 'helpadmin', 'helpowner'].includes(commandName)) {
       throw new Error(`the ${commandName} command is reserved`);
     }
     if (!commandModule || typeof commandModule.handler !== 'function') {
-      throw new Error('module handler must be of type function');
+      throw new Error(`module handler for ${commandName} must be of type function`);
     }
     if (commands[commandName]) {
       throw new Error(`Command ${commandName} already exists`);
@@ -44,13 +47,33 @@ for (const eachModuleFile of fs.readdirSync(commandDir)) {
   }
 }
 
+// initialize crontabs
+for (const eachModuleFile of fs.readdirSync(cronDir)) {
+  if (eachModuleFile.endsWith('.js') && eachModuleFile !== '.js') {
+    const cronModule = require(path.join(__dirname, cronDir, eachModuleFile));
+    const cronName = eachModuleFile.slice(0, -3);
+    if (!cronModule || typeof cronModule.handler !== 'function') {
+      throw new Error(`handler for cron module ${cronName} must be a function`);
+    }
+    if (!cronModule.crontab) {
+      throw new Error(`cron module for ${cronName} must have a crontab`);
+    }
+    cron.schedule(cronModule.crontab, () => cronModule.handler({bot, storage, firstRun: false}));
+    if (cronModule.runImmediately) cronsToRunImmediately.push(() => cronModule.handler({bot, storage, firstRun: true}));
+  }
+}
+
 bot.on('ready', () => {
   console.log(`Logged in as ${bot.user.tag}!`);
+  for (const eachCron of cronsToRunImmediately) eachCron();
 });
 
 function msgHandler(msg, recursiveCall = false) {
   const serverData = storage.getServerData(msg.guild.id);
   const prefix = serverData.prefix || defaultPrefix;
+
+  const isAdmin = () => msg.member.hasPermission('ADMINISTRATOR') || (msg.author.id === ownerId && viewOwnerAsAdmin);
+  const isOwner = () => msg.author.id === ownerId;
 
   // for debugging purposes
   if (msg.content === 'oh bot oh bot oh bot. kneel before me your divine creator.' && msg.author.id === ownerId) {
@@ -63,7 +86,17 @@ function msgHandler(msg, recursiveCall = false) {
     const command = splitCommand[0].slice(prefix.length);
     const args = splitCommand.slice(1);
     if (command === 'help') {
-      const commandsToList = Object.keys(commands).filter(cmd => !commands[cmd].adminOnly && !commands[cmd].ownerOnly);
+      const abilityToListAdminCommands = isAdmin() || isOwner();
+      const commandsToList = Object.keys(commands).filter(cmd => {
+        switch(abilityToListAdminCommands ? args[0] : '') {
+          case 'admin':
+            return commands[cmd].adminOnly;
+          case 'owner':
+            return commands[cmd].ownerOnly;
+          default:
+            return !commands[cmd].adminOnly && !commands[cmd].ownerOnly;
+        }
+      });
       const replyMsg = '**Commands**\n' + commandsToList.join(', ');
       let descriptionMsg = '';
       for (const cmdToList of commandsToList) {
@@ -72,8 +105,8 @@ function msgHandler(msg, recursiveCall = false) {
       return msg.lineReplyNoMention(replyMsg + (descriptionMsg ? '\n\n**Command descriptions**\n' + descriptionMsg : ''));
     }
     else if (commands[command] &&
-      (!commands[command].adminOnly || (msg.member.hasPermission('ADMINISTRATOR') || (msg.author.id === ownerId && viewOwnerAsAdmin))) &&
-      (!commands[command].ownerOnly || msg.author.id === ownerId)) {
+      (!commands[command].adminOnly || isAdmin()) &&
+      (!commands[command].ownerOnly || isOwner())) {
       commands[command].handler({msg, args, storage, bot});
     }
     else if (serverData.aliases && serverData.aliases[command]) {