+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v1.1.0
+
+- Add crontab functionality
+- Add tank of the week
+- Add ability to view current settings
+- Add help admin and help owner
+- Add eval for debugging purposes
+- Fix invalid number of arguments when unsetting prefix
+
+
 ## v1.0.0
 
 Initial commit